♻️ Extract ConnectionChooser with debounced search
- Move ConnectionChooser to its own file for better organization
- Add 300ms debounced search filter for performance
- Delete unused ConnectHeader component (layout now in connect-layout.tsx)
- Update exports in index.ts
