+++ components/connection/index.ts
@@ -2,5 +2,5 @@ export { Chat } from "./chat";
 export { ConciergeDisplay } from "./concierge-display";
 export { ConnectRuntimeProvider } from "./connect-runtime-provider";
 export { ConnectLayout } from "./connect-layout";
-export { ConnectHeader, ConnectionChooser } from "./connect-header";
+export { ConnectionChooser } from "./connection-chooser";
 export { ConnectionProvider, useConnection } from "./connection-context";
